// src/generated/models/notification-record-model.ts
import type { EmailTemplate } from './email-template-model';
import type { RequestedUser } from './requested-user-model';

export const NotificationRecordStatusKeyToLabel = {
  'Sent': 'Sent',
  'Failed': 'Failed',
  'Queued': 'Queued',
  'Sending': 'Sending'
} as const;
export type NotificationRecordStatusKey = keyof typeof NotificationRecordStatusKeyToLabel;

export interface NotificationRecord {
  /**
   * @displayName Notification Record
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Notification Record Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  notificationRecordName: string;
  /**
   * @displayName Attempt Number
   * @validationRule Required for create/update operations
   */
  attemptNumber: number;
  /**
   * @displayName BCC
   */
  bCC?: string;
  /**
   * @displayName Body Snapshot
   * @validationRule Required for create/update operations
   */
  bodySnapshot: string;
  /**
   * @displayName CC
   */
  cC?: string;
  /**
   * @displayName Email Template
   * @validationRule Required for create/update operations
   */
  emailTemplate: Pick<EmailTemplate, 'id' | 'emailTemplateName'>;
  /**
   * @displayName Error Details
   */
  errorDetails?: string;
  /**
   * @displayName Importance
   */
  importance?: boolean;
  /**
   * @displayName Last Updated Date
   */
  lastUpdatedDate?: string;
  /**
   * @displayName Message ID
   */
  messageID?: string;
  /**
   * @displayName Processing Date
   */
  processingDate?: string;
  /**
   * @displayName Recipient Email
   * @validationRule Required for create/update operations
   */
  recipientEmail: string;
  /**
   * @displayName Recipient Name
   */
  recipientName?: string;
  /**
   * @displayName Requested User
   * @validationRule Required for create/update operations
   */
  requestedUser: Pick<RequestedUser, 'id' | 'fullName'>;
  /**
   * @displayName Request Reference Snapshot
   * @validationRule Required for create/update operations
   */
  requestReferenceSnapshot: string;
  /**
   * @displayName Sent By
   */
  sentBy?: string;
  /**
   * @displayName Sent Date
   */
  sentDate?: string;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: NotificationRecordStatusKey;
  /**
   * @displayName Subject Snapshot
   * @validationRule Required for create/update operations
   */
  subjectSnapshot: string;
  /**
   * @displayName Support Contact Snapshot
   * @validationRule Required for create/update operations
   */
  supportContactSnapshot: string;
  /**
   * @displayName Template Version Snapshot
   * @validationRule Required for create/update operations
   */
  templateVersionSnapshot: number;
}

export const _NotificationRecord = 'NotificationRecord' as const;