// src/generated/models/provisioning-request-model.ts
export const ProvisioningRequestSourceTypeKeyToLabel = {
  'ServiceNow': 'ServiceNow',
  'Email': 'Email'
} as const;
export type ProvisioningRequestSourceTypeKey = keyof typeof ProvisioningRequestSourceTypeKeyToLabel;

export const ProvisioningRequestStatusKeyToLabel = {
  'New': 'New',
  'InReview': 'In Review',
  'InProgress': 'In Progress',
  'Completed': 'Completed',
  'Exception': 'Exception',
  'Cancelled': 'Cancelled',
  'OnHold': 'On Hold'
} as const;
export type ProvisioningRequestStatusKey = keyof typeof ProvisioningRequestStatusKeyToLabel;

export interface ProvisioningRequest {
  /**
   * @displayName Provisioning Request
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Provisioning Request Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  provisioningRequestName: string;
  /**
   * @displayName Cancellation Reason
   */
  cancellationReason?: string;
  /**
   * @displayName Cancelled Date
   */
  cancelledDate?: string;
  /**
   * @displayName Created Date
   * @validationRule Required for create/update operations
   */
  createdDate: string;
  /**
   * @displayName Exception Date
   */
  exceptionDate?: string;
  /**
   * @displayName Exception Reason
   */
  exceptionReason?: string;
  /**
   * @displayName Hold Date
   */
  holdDate?: string;
  /**
   * @displayName Hold Reason
   */
  holdReason?: string;
  /**
   * @displayName Notes
   */
  notes?: string;
  /**
   * @displayName Owner Email
   * @validationRule Required for create/update operations
   */
  ownerEmail: string;
  /**
   * @displayName Owner Name
   * @validationRule Required for create/update operations
   */
  ownerName: string;
  /**
   * @displayName Source Reference
   * @validationRule Required for create/update operations
   */
  sourceReference: string;
  /**
   * @displayName Source Type
   * @validationRule Required for create/update operations
   */
  sourceTypeKey: ProvisioningRequestSourceTypeKey;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: ProvisioningRequestStatusKey;
  /**
   * @displayName Summary
   * @validationRule Required for create/update operations
   */
  summary: string;
}

export const _ProvisioningRequest = 'ProvisioningRequest' as const;