// src/generated/models/requested-user-stage-progress-model.ts
import type { RequestUser } from './request-user-model';
import type { WorkflowStageOption } from './workflow-stage-option-model';

export const RequestedUserStageProgressStateKeyToLabel = {
  'NotStarted': 'Not Started',
  'InProgress': 'In Progress',
  'Completed': 'Completed',
  'Blocked': 'Blocked',
  'Skipped': 'Skipped'
} as const;
export type RequestedUserStageProgressStateKey = keyof typeof RequestedUserStageProgressStateKeyToLabel;

export interface RequestedUserStageProgress {
  /**
   * @displayName Requested User Stage Progress
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Stage Progress Name
   * @description Human-readable record identifier
   */
  stageProgressName: string;
  /**
   * @displayName Request User
   */
  requestUser?: Pick<RequestUser, 'id' | 'requestUserName'>;
  /**
   * @displayName Sequence Number
   */
  sequenceNumber?: number;
  /**
   * @displayName State
   */
  stateKey?: RequestedUserStageProgressStateKey;
  /**
   * @displayName Workflow Stage Option
   */
  workflowStageOption?: Pick<WorkflowStageOption, 'id' | 'stageName'>;
}

export const _RequestedUserStageProgress = 'RequestedUserStageProgress' as const;